--- mobile/src/features/Buyer/index.ts
@@ -1,2 +1,0 @@
-export * from './ui/SellerDashboard';
-// export * from './model/store';
