--- mobile/src/features/Buyer/ui/styles.ts
@@ -1,41 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-import {Colors} from 'shared/libs/helpers/colors';
-
-export const LoginStyles = StyleSheet.create({
-  container: {
-    flex: 1,
-    paddingTop: 60,
-    backgroundColor: Colors.White100,
-  },
-  textCenter: {
-    textAlign: 'center',
-    marginBottom: 20,
-  },
-  image: {
-    width: 150,
-    height: 150,
-    borderRadius: 75,
-    marginBottom: 20,
-    marginLeft: 'auto',
-    marginRight: 'auto',
-  },
-  menuContainer: {
-    marginTop: 20,
-    paddingHorizontal: 20,
-  },
-  menuItem: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingVertical: 15,
-    borderBottomWidth: 1,
-    borderBottomColor: Colors.Gray400,
-  },
-  menuItemText: {
-    flex: 1,
-    marginLeft: 15,
-    fontSize: 16,
-    color: Colors.Black100,
-  },
-});
