--- mobile/src/features/Seller/index.ts
@@ -1,2 +1,0 @@
-export * from './ui/Seller';
-// export * from './model/store';
