+++ mobile/src/features/index.ts
@@ -2,5 +2,4 @@ export * from './Login';
 export * from './Register';
 export * from './ProductInfo';
 export * from './Search';
-export * from './Seller';
 export * from './ProductListing';
