+++ mobile/src/pages/Profile/ui/Profile.tsx
@@ -1,103 +1,113 @@
-import React from 'react';
-import {View} from 'react-native';
+import {NavigationProp, useNavigation} from '@react-navigation/native';
+import {IndexPath, Menu, MenuItem} from '@ui-kitten/components';
+import {useState} from 'react';
+import {useTranslation} from 'react-i18next';
+import {Image, Text, View} from 'react-native';
 
+import {Screens} from 'app/navigation/navigationEnums';
+import {ProfileStackParamsList} from 'app/navigation/navigationTypes';
 import {useUserStore} from 'entities/user';
-import {Seller} from 'features';
-import {Auth} from 'pages';
+import {Auth} from 'pages/index';
+import {
+  FlagIcon,
+  ForwardIcon,
+  HistoryIcon,
+  LogOutIcon,
+  ProductIcon,
+} from 'shared/icons';
+import {Colors, TextStyles} from 'shared/libs/helpers';
+
+import {ProfileStyles as styles} from './Profile.styles';
 
 export const Profile = () => {
   const {token, user, clearUserData} = useUserStore();
+  const [selectedIndex, setSelectedIndex] = useState<IndexPath | undefined>(
+    undefined,
+  );
+  const {t} = useTranslation();
+
+  const navigation = useNavigation<NavigationProp<ProfileStackParamsList>>();
 
   const handleLogout = () => {
     clearUserData();
   };
 
+  const handlePurchaseHistory = () => {
+    navigation.navigate(Screens.ORDERS);
+  };
+
+  const handleSettings = () => {
+    navigation.navigate(Screens.SETTINGS);
+  };
+
+  const handleAddProduct = () => {
+    navigation.navigate(Screens.NEWPRODUCT);
+  };
+
+  const renderSellerInterface = () => (
+    <MenuItem
+      title={t('Добавить продукт')}
+      accessoryLeft={ProductIcon}
+      accessoryRight={ForwardIcon}
+      onPress={handleAddProduct}
+    />
+  );
+
+  const renderBuyerInterface = () => (
+    <MenuItem
+      title={t('История покупок')}
+      accessoryLeft={HistoryIcon}
+      accessoryRight={ForwardIcon}
+      onPress={handlePurchaseHistory}
+    />
+  );
+
+  const isSeller = user.role === 'seller';
+
   return (
     <>
       {token ? (
-        user.role === 'seller' ? (
-          <Seller user={user} onLogout={handleLogout} />
-        ) : (
-          <View></View>
-          // <Buyer onLogout={handleLogout} />
-        )
+        <View style={styles.container}>
+          <Text
+            style={[
+              TextStyles.h2.changeColor(Colors.Black200),
+              styles.textCenter,
+            ]}>
+            {t('Профиль')}
+          </Text>
+          <Image
+            style={styles.image}
+            source={require('shared/assets/images/tempimage.png')}
+          />
+          <Text
+            style={[
+              TextStyles.p2.changeColor(Colors.Black200),
+              styles.textCenter,
+            ]}>
+            {`${user.name} ${user.surname}`}
+          </Text>
+
+          <View style={styles.menuContainer}>
+            <Menu onSelect={index => setSelectedIndex(index)}>
+              {isSeller ? renderSellerInterface() : renderBuyerInterface()}
+              <MenuItem
+                title={t('Настройки языка')}
+                accessoryLeft={FlagIcon}
+                accessoryRight={ForwardIcon}
+                onPress={handleSettings}
+              />
+              <MenuItem
+                title={t('Выйти из аккаунта')}
+                accessoryLeft={LogOutIcon}
+                accessoryRight={ForwardIcon}
+                onPress={handleLogout}
+              />
+            </Menu>
+          </View>
+        </View>
       ) : (
         <Auth />
       )}
     </>
   );
 };
-
-// export const Profile = () => {
-//   const {token, user, clearUserData} = useUserStore();
-//   const [selectedIndex, setSelectedIndex] = useState<IndexPath | undefined>(
-//     undefined,
-//   );
-//   const {t} = useTranslation();
-
-//   const navigation = useNavigation<NavigationProp<ProfileStackParamsList>>();
-
-//   const handlePurchaseHistory = () => {
-//     navigation.navigate(Screens.ORDERS);
-//   };
-
-//   const handleSettings = () => {
-//     navigation.navigate(Screens.SETTINGS);
-//   };
-
-//   const handleLogout = () => {
-//     clearUserData();
-//   };
-
-//   return (
-//     <>
-//       {token ? (
-//         <View style={styles.container}>
-//           <Text
-//             style={[
-//               TextStyles.h2.changeColor(Colors.Black200),
-//               styles.textCenter,
-//             ]}>
-//             {t('Профиль')}
-//           </Text>
-//           <Image
-//             style={styles.image}
-//             source={require('shared/assets/images/tempimage.png')}
-//           />
-//           <Text
-//             style={[
-//               TextStyles.p2.changeColor(Colors.Black200),
-//               styles.textCenter,
-//             ]}>
-//             {`${user.name} ${user.surname}`}
-//           </Text>
-
-//           <View style={styles.menuContainer}>
-//             <Menu onSelect={index => setSelectedIndex(index)}>
-//               <MenuItem
-//                 title={t('История покупок')}
-//                 accessoryLeft={HistoryIcon}
-//                 accessoryRight={ForwardIcon}
-//                 onPress={handlePurchaseHistory}
-//               />
-//               <MenuItem
-//                 title={t('Настройки языка')}
-//                 accessoryLeft={FlagIcon}
-//                 accessoryRight={ForwardIcon}
-//                 onPress={handleSettings}
-//               />
-//               <MenuItem
-//                 title={t('Выйти из аккаунта')}
-//                 accessoryLeft={LogOutIcon}
-//                 accessoryRight={ForwardIcon}
-//                 onPress={handleLogout}
-//               />
-//             </Menu>
-//           </View>
-//         </View>
-//       ) : (
-//         <Auth />
-//       )}
-//     </>
-//   );
-// };
